requirements.txt
streamlit
sqlalchemy
pandas
langchain-groq
python-dotenv
mysql-connector-python